Added snackbar to the settings form
